requirements.txt
numpy==1.26.2
onnxruntime==1.16.3
pandas==2.1.4
Pillow==10.1.0
PyQt6==6.6.1
